+++ TacoBell/Models/BusinessLogicLayer/DishBLL.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using TacoBell.Models.Entities;
 
@@ -61,14 +62,33 @@ namespace TacoBell.Models.BusinessLogicLayer
             }
         }
 
-        public void DeleteDish(int id)
+        public void DeleteDish(int dishId)
         {
-            var dish = _db.Dishes.Find(id);
+            var dish = _db.Dishes.Find(dishId);
             if (dish != null)
             {
+                bool inMenus = _db.MenuDishes.Any(md => md.DishId == dishId);
+                bool inOrders = _db.OrderDishes.Any(od => od.DishId == dishId);
+
+                if (inMenus)
+                    throw new InvalidOperationException("Acest preparat este inclus într-un meniu și nu poate fi șters.");
+                if (inOrders)
+                    throw new InvalidOperationException("Acest preparat este inclus într-o comandă și nu poate fi șters.");
+
+                // Legături care nu blochează integritatea
+                var allergens = _db.DishAllergens.Where(da => da.DishId == dishId).ToList();
+                if (allergens.Any()) _db.DishAllergens.RemoveRange(allergens);
+
+                var images = _db.DishImages.Where(img => img.DishId == dishId).ToList();
+                if (images.Any()) _db.DishImages.RemoveRange(images);
+
                 _db.Dishes.Remove(dish);
                 _db.SaveChanges();
             }
         }
+
+
+
+
     }
 }
